DRY up custom types trick, so it will only be in our bundle output once
